requirements.txt
fastapi
uvicorn
python-nmap
requests
jinja2
openai==0.28.0